[#17] 매장 관리자를 등록하는 어드민 기능 구현

+++ src/main/java/dabang/star/cafe/api/OfficeAdminApi.java
@@ -61,7 +61,7 @@ public class OfficeAdminApi {
      */
     @LoginCheck(role = Role.ADMIN)
     @DeleteMapping("/{officeId}")
-    public void deleteOffice(@PathVariable Long officeId) {
+    public void deleteOffice(@PathVariable int officeId) {
         officeAdminService.deleteOffice(officeId);
     }
