+++ src/main/java/dabang/star/cafe/domain/admin/OfficeAdminService.java
@@ -10,9 +10,9 @@ public interface OfficeAdminService {
 
     OfficeData createOffice(String name, String address, BigDecimal latitude, BigDecimal longitude);
 
-    void deleteOffice(Long officeId);
+    void deleteOffice(int officeId);
 
-    void updateOffice(Long id, String name, String address, BigDecimal latitude, BigDecimal longitude);
+    void updateOffice(int id, String name, String address, BigDecimal latitude, BigDecimal longitude);
 
     List<OfficeData> findOffices(Page page);
 }
